+++ resources/assets/components/actions/TextSubmissionAction/TextSubmissionAction.js
@@ -6,9 +6,9 @@ import PropTypes from 'prop-types';
 import classnames from 'classnames';
 
 import Markdown from '../../Markdown';
-import Button from '../../Button/Button';
 import Card from '../../utilities/Card/Card';
 import Modal from '../../utilities/Modal/Modal';
+import Button from '../../utilities/Button/Button';
 import { getFieldErrors } from '../../../helpers/forms';
 import FormValidation from '../../utilities/Form/FormValidation';
