+++ resources/assets/components/utilities/Button/button.scss
@@ -1,4 +1,4 @@
-@import '../../scss/next-toolbox.scss';
+@import '../../../scss/next-toolbox.scss';
 
 .button {
   &.-inline {
