+++ resources/assets/components/utilities/LoadMore/LoadMore.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classnames from 'classnames';
 
-import Button from '../../Button/Button';
+import Button from '../../utilities/Button/Button';
 
 const LoadMore = ({
   onClick,
